Rename package for appender
