+++ uiautomationutilities/src/main/java/com/microsoft/identity/client/ui/automation/logging/AppenderRegistry.java
@@ -24,7 +24,7 @@ package com.microsoft.identity.client.ui.automation.logging;
 
 import androidx.annotation.NonNull;
 
-import com.microsoft.identity.client.ui.automation.logging.appenders.IAppender;
+import com.microsoft.identity.client.ui.automation.logging.appender.IAppender;
 
 import java.util.HashSet;
 import java.util.Set;
