+++ uiautomationutilities/src/main/java/com/microsoft/identity/client/ui/automation/logging/appender/LogcatAppender.java
@@ -20,12 +20,11 @@
 //  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 //  THE SOFTWARE.
-package com.microsoft.identity.client.ui.automation.logging.appenders;
+package com.microsoft.identity.client.ui.automation.logging.appender;
 
 import android.util.Log;
 
 import com.microsoft.identity.client.ui.automation.logging.LogLevel;
-import com.microsoft.identity.client.ui.automation.logging.appenders.IAppender;
 
 /**
  * An implementation of {@link IAppender} to send logs to Android logcat.
